+++ packages/react/src/components/carbon-components-react/UIShell/HeaderMenu.js
@@ -81,6 +81,7 @@ class HeaderMenu extends React.Component {
       selectedIndex: null,
     };
     this.items = [];
+    this.menuLinkRef = React.createRef();
   }
 
   /**
@@ -125,6 +126,7 @@ class HeaderMenu extends React.Component {
       `${prefix}--masthead__megamenu__category-headline`,
       `${prefix}--masthead__megamenu__category-group`,
       `${prefix}--masthead__megamenu__view-all-cta`,
+      `${prefix}--masthead__megamenu__l0-nav`,
     ];
 
     return megamenuItems.filter(item =>
@@ -192,6 +194,14 @@ class HeaderMenu extends React.Component {
     }
   };
 
+  /**
+   * This forces focus on the menu link when clicked on
+   */
+  handleFocus = event => {
+    event.preventDefault();
+    this.menuLinkRef.current.focus();
+  };
+
   render() {
     const {
       'aria-label': ariaLabel,
@@ -227,11 +237,12 @@ class HeaderMenu extends React.Component {
           aria-expanded={this.state.expanded}
           className={`${prefix}--header__menu-item ${prefix}--header__menu-title`}
           href="#"
-          onClick={event => event.preventDefault()}
+          onClick={this.handleFocus}
           onKeyDown={this.handleOnKeyDown}
           ref={this.handleMenuButtonRef}
           role="menuitem"
           tabIndex={0}
+          ref={this.menuLinkRef}
           {...accessibilityLabel}>
           {menuLinkName}
           <MenuContent />