+++ packages/services/src/services/Locale/__tests__/Locale.test.js
@@ -4,7 +4,6 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-
 import { geolocation, ipcinfoCookie } from '@carbon/ibmdotcom-utilities';
 import digitalDataResponse from '../../DDO/__tests__/data/response.json';
 import LocaleAPI from '../Locale';
@@ -17,8 +16,8 @@ const mockDigitalDataResponse = digitalDataResponse;
 jest.mock(
   '@carbon/ibmdotcom-utilities/lib/utilities/ipcinfoCookie/ipcinfoCookie',
   () => ({
-    get: jest.fn(() => Promise.resolve({ cc: 'us', lc: 'en' })),
-    set: jest.fn(() => Promise.resolve({})),
+    get: jest.fn(() => ({ cc: 'us', lc: 'en' })),
+    set: jest.fn(() => ({})),
   })
 );
 
@@ -28,27 +27,16 @@ jest.mock(
 );
 
 describe('LocaleAPI', () => {
-  const _cc = 'us';
-  const _lc = 'en';
-
-  const endpoint = `${process.env.REACT_APP_CORS_PROXY}${process.env.TRANSLATION_HOST}/common/js/dynamicnav/www/countrylist/jsononly`;
-  const fetchUrl = `${endpoint}/${_cc}${_lc}-utf8.json`;
+  const handles = [];
 
   beforeEach(function() {
-    mockAxios.get.mockImplementation(() =>
-      Promise.resolve({
-        data: response,
-      })
-    );
+    mockAxios.get.mockImplementation(async () => ({
+      data: response,
+    }));
 
     root.digitalData = mockDigitalDataResponse;
 
     LocaleAPI.clearCache();
-  });
-
-  afterEach(() => {
-    jest.clearAllMocks();
-    sessionStorage.clear();
   });
 
   it('should fetch the lang from the html attribute', async () => {
@@ -88,51 +76,421 @@ describe('LocaleAPI', () => {
     });
   });
 
-  it('should fetch the display name based on language/locale combination', async () => {
-    const data = await LocaleAPI.getLangDisplay();
-    expect(data).toEqual('United States — English');
+  it('should default when ddo is undefined', async () => {
+    root.digitalData = undefined;
+    const lang = await LocaleAPI.getLang();
+
+    expect(lang).toEqual({
+      cc: 'us',
+      lc: 'en',
+    });
   });
 
-  it('should fetch locale from cookie if availiable', async () => {
-    Object.defineProperty(window.document, 'cookie', {
-      writable: true,
-      value: 'ipcInfo=cc%253Dus%253Blc%253Den',
+  it('should default when no ddo.page', async () => {
+    root.digitalData.page = false;
+    const lang = await LocaleAPI.getLang();
+
+    expect(lang).toEqual({
+      cc: 'us',
+      lc: 'en',
     });
+  });
+
+  it('should default when no ddo.page.pageInfo', async () => {
+    root.digitalData = {
+      page: {
+        pageInfo: false,
+      },
+    };
+    const lang = await LocaleAPI.getLang();
+
+    expect(lang).toEqual({
+      cc: 'us',
+      lc: 'en',
+    });
+  });
+
+  it('should default when no ddo.page.pageInfo.ibm', async () => {
+    root.digitalData = {
+      page: {
+        pageInfo: {
+          ibm: false,
+        },
+      },
+    };
+    const lang = await LocaleAPI.getLang();
+
+    expect(lang).toEqual({
+      cc: 'us',
+      lc: 'en',
+    });
+  });
+
+  it('should default when no ddo.page.pageInfo.ibm.country', async () => {
+    root.digitalData = {
+      page: {
+        pageInfo: {
+          ibm: {
+            country: false,
+          },
+        },
+      },
+    };
+    const lang = await LocaleAPI.getLang();
+
+    expect(lang).toEqual({
+      cc: 'us',
+      lc: 'en',
+    });
+  });
+
+  it('should default when no ddo.page.pageInfo.language', async () => {
+    root.digitalData = {
+      page: {
+        pageInfo: {
+          language: false,
+          ibm: {
+            country: 'de',
+          },
+        },
+      },
+    };
+    const lang = await LocaleAPI.getLang();
+
+    expect(lang).toEqual({
+      cc: 'us',
+      lc: 'en',
+    });
+  });
+
+  it('should get from DDO', async () => {
+    root.digitalData = {
+      page: {
+        pageInfo: {
+          language: 'de-DE',
+          ibm: {
+            country: 'de',
+          },
+        },
+      },
+    };
+    const lang = await LocaleAPI.getLang();
+
+    expect(lang).toEqual({
+      cc: 'de',
+      lc: 'de',
+    });
+  });
+
+  it('should handle multiple countries', async () => {
+    root.digitalData = {
+      page: {
+        pageInfo: {
+          language: 'de-DE',
+          ibm: {
+            country: 'de,ca',
+          },
+        },
+      },
+    };
+    const lang = await LocaleAPI.getLang();
+
+    expect(lang).toEqual({
+      cc: 'de',
+      lc: 'de',
+    });
+  });
+
+  it('should handle map gb to uk', async () => {
+    root.digitalData = {
+      page: {
+        pageInfo: {
+          language: 'de-DE',
+          ibm: {
+            country: 'gb',
+          },
+        },
+      },
+    };
+    const lang = await LocaleAPI.getLang();
+
+    expect(lang).toEqual({
+      cc: 'uk',
+      lc: 'de',
+    });
+  });
+
+  it('should handle map zz to us', async () => {
+    root.digitalData = {
+      page: {
+        pageInfo: {
+          language: 'de-DE',
+          ibm: {
+            country: 'zz',
+          },
+        },
+      },
+    };
+    const lang = await LocaleAPI.getLang();
+
+    expect(lang).toEqual({
+      cc: 'us',
+      lc: 'de',
+    });
+  });
+
+  it('should get countries list', async () => {
+    const countries = await LocaleAPI.getList({
+      cc: 'testCC',
+      lc: 'testLC',
+    });
+
+    expect(mockAxios.get).toHaveBeenCalledTimes(1);
+    expect(
+      mockAxios.get
+    ).toHaveBeenCalledWith(
+      'https://cra-myproxy.com/https://ibm.com/common/js/dynamicnav/www/countrylist/jsononly/testCCtestLC-utf8.json',
+      { headers: { 'Content-Type': 'application/json; charset=utf-8' } }
+    );
+
+    expect(countries).toEqual(response);
+  });
+
+  it('should get countries list from session cache', async () => {
+    mockAxios.get.mockClear();
+    const countries1 = await LocaleAPI.getList({
+      cc: 'testCC',
+      lc: 'testLC',
+    });
+    const countries2 = await LocaleAPI.getList({
+      cc: 'testCC',
+      lc: 'testLC',
+    });
+
+    expect(mockAxios.get).toHaveBeenCalledTimes(1);
+    expect(countries1).toEqual(response);
+    expect(countries2).toEqual(response);
+  });
+
+  it('should get default countries list on inital reject', async () => {
+    mockAxios.get.mockClear();
+    mockAxios.get
+      .mockReturnValueOnce(Promise.reject())
+      .mockReturnValueOnce(Promise.resolve({ data: response }));
+
+    const countries = await LocaleAPI.getList({
+      cc: 'testCC',
+      lc: 'testLC',
+    });
+
+    expect(mockAxios.get).toHaveBeenCalledTimes(2);
+    expect(
+      mockAxios.get
+    ).toHaveBeenCalledWith(
+      'https://cra-myproxy.com/https://ibm.com/common/js/dynamicnav/www/countrylist/jsononly/testCCtestLC-utf8.json',
+      { headers: { 'Content-Type': 'application/json; charset=utf-8' } }
+    );
+    expect(
+      mockAxios.get
+    ).toHaveBeenCalledWith(
+      'https://cra-myproxy.com/https://ibm.com/common/js/dynamicnav/www/countrylist/jsononly/usen-utf8.json',
+      { headers: { 'Content-Type': 'application/json; charset=utf-8' } }
+    );
+
+    expect(countries).toEqual(response);
+  });
+
+  it('should reject countries list', async () => {
+    mockAxios.get.mockImplementation(() => Promise.reject());
+
+    let caught;
+    try {
+      await LocaleAPI.getList({ cc: 'us', lc: 'en' });
+    } catch (error) {
+      caught = true;
+    }
+    expect(caught).not.toBeNull();
+  });
+
+  it('should verify locale', () => {
+    const locale = LocaleAPI.verifyLocale('us', 'en', {
+      regionList: [
+        {
+          countryList: [
+            {
+              locale: [['en-us']],
+            },
+            {
+              locale: [['de-de']],
+            },
+          ],
+        },
+      ],
+    });
+
+    expect(locale).toEqual({
+      cc: 'us',
+      lc: 'en',
+    });
+  });
+
+  it('should verify priority locale', () => {
+    const locale = LocaleAPI.verifyLocale('us', 'fr', {
+      regionList: [
+        {
+          countryList: [
+            {
+              locale: [['en-us']],
+            },
+            {
+              locale: [['de-de']],
+            },
+          ],
+        },
+      ],
+    });
+
+    expect(locale).toEqual({
+      cc: 'us',
+      lc: 'en',
+    });
+  });
+
+  it('should verify undefined', () => {
+    const locale = LocaleAPI.verifyLocale('us', 'en', false);
+
+    expect(locale).toBeUndefined();
+  });
+
+  it('should get lang display', async () => {
+    jest.spyOn(LocaleAPI, 'getLang').mockReturnValue(
+      Promise.resolve({
+        cc: 'testCC',
+        lc: 'testLC',
+      })
+    );
+    jest.spyOn(LocaleAPI, 'getList').mockReturnValue(
+      Promise.resolve({
+        regionList: [
+          {
+            countryList: [
+              {
+                name: 'testName',
+                locale: [
+                  ['testLC-testCC', 'testDisplay'],
+                  ['altLocale', 'altDisplay'],
+                ],
+              },
+              {
+                locale: [['testLC2-testCC2', 'testDisplay2']],
+              },
+            ],
+          },
+        ],
+      })
+    );
+
+    const display = await LocaleAPI.getLangDisplay({
+      cc: 'testCC',
+      lc: 'testLC',
+    });
+
+    expect(display).toEqual('testName — testDisplay');
+  });
+
+  it('should get lang for display', async () => {
+    const display = await LocaleAPI.getLangDisplay(false);
+
+    expect(display).toEqual('testName — testDisplay');
+  });
+
+  it('should get default lang display', async () => {
+    const display = await LocaleAPI.getLangDisplay({
+      cc: 'missingCC',
+      lc: 'missingLC',
+    });
+
+    expect(display).toEqual('United States — English');
+  });
+
+  it('should get locale from getLang', async () => {
+    jest
+      .spyOn(LocaleAPI, 'getLang')
+      .mockReturnValue(Promise.resolve('testLang'));
+    jest
+      .spyOn(LocaleAPI, 'getList')
+      .mockReturnValue(Promise.resolve('testList'));
+    jest.spyOn(LocaleAPI, 'verifyLocale').mockReturnValue('testVerified');
+    const locale = await LocaleAPI.getLocale();
+
+    expect(locale).toEqual('testLang');
+  });
+
+  it('should get locale from cookies', async () => {
+    jest.spyOn(LocaleAPI, 'getLang').mockReturnValue(Promise.resolve(false));
 
     const locale = await LocaleAPI.getLocale();
 
     expect(locale).toEqual({ cc: 'us', lc: 'en' });
   });
 
-  it('should verify the locale and return existing combo', () => {
-    const locale = LocaleAPI.verifyLocale('us', 'es', response);
+  it('should get locale from geolocation on missing cookie', async () => {
+    ipcinfoCookie.get.mockImplementation(() => false);
 
-    expect(locale).toEqual({ cc: 'us', lc: 'en' });
-  });
-
-  xit('should set the ipcInfo cookie once combo has been verified', async () => {
     await LocaleAPI.getLocale();
 
-    expect(ipcinfoCookie.set).toHaveBeenCalledTimes(1);
     expect(geolocation).toHaveBeenCalledTimes(1);
   });
 
-  it('should make the call for the country list', async () => {
-    const data = await LocaleAPI.getList({ cc: _cc, lc: _lc });
+  it('should get locale from geolocation on missing cookie lc', async () => {
+    geolocation.mockClear();
+    ipcinfoCookie.get.mockImplementation(() => ({ cc: 'testCC' }));
 
-    expect(data).toEqual(response);
-    expect(mockAxios.get).toHaveBeenCalledWith(fetchUrl, {
-      headers: {
-        'Content-Type': 'application/json; charset=utf-8',
-      },
-    });
+    await LocaleAPI.getLocale();
+
+    expect(geolocation).toHaveBeenCalledTimes(1);
+  });
+
+  it('should get locale from geolocation on missing cookie cc', async () => {
+    geolocation.mockClear();
+    ipcinfoCookie.get.mockImplementation(() => ({ lc: 'testLC' }));
+
+    await LocaleAPI.getLocale();
+
+    expect(geolocation).toHaveBeenCalledTimes(1);
+  });
+
+  it('should get locale from geolocation', async () => {
+    ipcinfoCookie.set.mockClear();
+    ipcinfoCookie.get.mockImplementation(() => false);
+
+    const locale = await LocaleAPI.getLocale();
+
+    expect(locale).toEqual('testVerified');
+    expect(ipcinfoCookie.set).toHaveBeenCalledTimes(1);
+    expect(ipcinfoCookie.set).toHaveBeenCalledWith('testVerified');
+  });
+
+  it('should get undefined locale on no cc', async () => {
+    geolocation.mockImplementation(() => Promise.resolve(false));
+
+    const locale = await LocaleAPI.getLocale();
+    expect(locale).toBeUndefined();
   });
 
   it('should use the cache for the country list, keyed by locale', async () => {
+    mockAxios.get.mockClear();
+    LocaleAPI.getList.mockRestore();
     await LocaleAPI.getList({ cc: 'us', lc: 'en' });
     await LocaleAPI.getList({ cc: 'us', lc: 'en' });
     await LocaleAPI.getList({ cc: 'kr', lc: 'ko' });
 
     expect(mockAxios.get).toHaveBeenCalledTimes(2);
+  });
+
+  afterEach(() => {
+    for (let handle = handles.pop(); handle; handle = handles.pop()) {
+      handle.release();
+    }
   });
 });